darchman/requirements/common.txt
flask
flask-restx
requests
werkzeug
pyyaml
gunicorn